Cambiado ciertos tipos de las entidades y nombres de columnas. Creacion de eventController

+++ src/main/java/com/Clubbr/Clubbr/Entity/stablishment.java
@@ -18,7 +18,7 @@ public class stablishment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "stablishmentID")
-    private int stablishmentID;
+    private Long stablishmentID;
 
     @Column(name = "stabName")
     private String stabName;
